fix(network.go): fix tests for network
The types of some of the properties of the overpass structures have changed

+++ overpass/network_test.go
@@ -42,8 +42,8 @@ func TestExecuteQuery(t *testing.T) {
 		Elements: []*Way{
 			{
 				Id:     1,
-				Bounds: [4]int{2, 3, 4, 5},
-				Nodes:  []int{6, 7},
+				Bounds: [4]float64{2, 3, 4, 5},
+				Nodes:  []uint64{6, 7},
 				Geometry: []*LatLon{
 					{8.0, 9.0},
 				},
